refactor: restaurantservice 패키지 분리
