+++ src/main/java/com/michelet/restaurantservice/checkin/presentation/controller/external/RestaurantCheckInController.java
@@ -8,8 +8,8 @@ import com.michelet.common.response.ApiResponse;
 import com.michelet.restaurantservice.checkin.application.command.CheckInCommand;
 import com.michelet.restaurantservice.checkin.application.result.CheckInResult;
 import com.michelet.restaurantservice.checkin.application.service.RestaurantCheckInCommandService;
-import com.michelet.restaurantservice.domain.exception.RestaurantErrorCode;
-import com.michelet.restaurantservice.domain.exception.RestaurantException;
+import com.michelet.restaurantservice.restaurant.domain.exception.RestaurantErrorCode;
+import com.michelet.restaurantservice.restaurant.domain.exception.RestaurantException;
 import com.michelet.restaurantservice.checkin.presentation.dto.CheckInResponse;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
