--- src/test/java/com/michelet/restaurant/MicheletApplicationTests.java
@@ -1,13 +1,0 @@
-package com.michelet.restaurant;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MicheletApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
