+++ src/test/java/com/michelet/restaurantservice/MicheletApplicationTests.java
@@ -1,0 +1,13 @@
+package com.michelet.restaurantservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class MicheletApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
