+++ src/test/java/com/michelet/restaurantservice/health/presentation/HealthControllerTest.java
@@ -1,0 +1,41 @@
+package com.michelet.restaurantservice.health.presentation;
+
+import static org.mockito.BDDMockito.given;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+import com.michelet.restaurantservice.health.application.HealthCheckService;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
+import org.springframework.test.web.servlet.MockMvc;
+
+/**
+ * 헬스체크 컨트롤러 테스트
+ *
+ * restaurant-service 헬스체크 엔드포인트의 정상 응답을 검증
+ */
+@WebMvcTest(HealthController.class)
+class HealthControllerTest {
+
+    @Autowired
+    private MockMvc mockMvc;
+
+    @MockitoBean
+    private HealthCheckService healthCheckService;
+
+    @Test
+    @DisplayName("헬스 체크")
+    void 헬스체크() throws Exception {
+        given(healthCheckService.getHealthStatus())
+                .willReturn("Restaurant Query Service is Healthy");
+
+        mockMvc.perform(get("/api/v1/restaurants/health"))
+                .andExpect(status().isOk())
+                .andExpect(jsonPath("$.success").value(true))
+                .andExpect(jsonPath("$.data").value("Restaurant Query Service is Healthy"));
+    }
+}
